--- app/src/main/java/com/example/ingredient/src/IngredientService.kt
@@ -1,5 +1,0 @@
-package com.example.ingredient.src
-
-class IngredientService {
-
-}
