--- app/src/main/java/com/example/ingredient/src/IngredientView.kt
@@ -1,5 +1,0 @@
-package com.example.ingredient.src
-
-interface IngredientView {
-
-}
