+++ app/src/main/java/com/example/ingredient/src/expirationDate/IngredientService.kt
@@ -1,0 +1,5 @@
+package com.example.ingredient.src.expirationDate
+
+class IngredientService {
+
+}
